Get commits associated chronologically with tags

Get commits associated chronologically with tags

get_commits_tags(repo = ".", ref = "master", path = NULL, silent = FALSE)

Arguments

repo
a path to a repository or a git_repository object. Default is '.'

ref
The name of a reference to list commits from e.g. a tag or a branch. The default is NULL for the current branch.

path
The path to a file. If not NULL, only commits modifying this file will be returned. Note that modifying commits that occurred before the file was given its present name are not returned; that is, the output of git log with --no-follow is reproduced.

silent
Logical. Whether to hide messages.

Value

A tibble with commits and tags

Examples

repo <- fake_repo()
get_commits_tags(repo = repo)

#> 4 commits found.

#> # A tibble: 4 x 9
#>   sha   summary message author email when                order tag.name
#>   <chr> <chr>   <chr>   <chr>  <chr> <dttm>              <int> <chr>   
#> 1 ca34… Add NE… "Add N… Alice  alic… 2020-09-09 13:23:40     4 NA      
#> 2 a3ad… Third … "Third… Alice  alic… 2020-09-09 13:23:40     3 v0.1    
#> 3 d387… exampl… "examp… Alice  alic… 2020-09-09 13:23:40     2 v0.1    
#> 4 a295… First … "First… Alice  alic… 2020-09-09 13:23:40     1 v0.1    
#> # … with 1 more variable: tag.message <chr>