Creates and updates a vignette in the 'vignette/' directory of a package with last modification time of tracked files

Creates and updates a vignette in the 'vignette/' directory of a package with last modification time of tracked files

create_vignette_last_modif(
  repo = ".",
  path = "R",
  recursive = TRUE,
  untracked = TRUE
)

update_vignette_last_modif(
  repo = ".",
  path = "R",
  recursive = TRUE,
  untracked = TRUE
)

Arguments

repo
git repository

path
Default to R folder. Use "" for the complete repository.

recursive
Logical. Should the listing recurse into directories?

untracked
Logical. Should the not tracked files be included?

Value

Creates and updates the content of the "modification_files.Rmd" vignette

Examples

# Creates vignette
repo <- fake_repo(as.package = TRUE)
create_vignette_last_modif(repo, path = "R")
# update vignette
repo <- fake_repo(as.package = TRUE)
update_vignette_last_modif(repo, path = "R")
rmarkdown::render(file.path(repo, "vignettes", "modification_files.Rmd"))

#> 
#> 
#> processing file: modification_files.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |.......................                                               |  33%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................................                       |  67%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: modification_files.knit.md

#> Warning: The vignette title specified in \VignetteIndexEntry{} is different from the title in the YAML metadata. The former is "zz-modification-of-files", and the latter is "Modifications of files". If that is intentional, you may set options(rmarkdown.html_vignette.check_title = FALSE) to suppress this check.

#> /usr/local/bin/pandoc +RTS -K512m -RTS modification_files.utf8.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output modification_files.html --lua-filter /Users/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /Users/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/latex-div.lua --self-contained --standalone --section-divs --template /Users/runner/work/_temp/Library/rmarkdown/rmd/h/default.html --highlight-style pygments --css /Users/runner/work/_temp/Library/rmarkdown/rmarkdown/templates/html_vignette/resources/vignette.css --include-in-header /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmprSGgWe/rmarkdown-str7a751fb6698.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' 

#> 
#> Output created: modification_files.html