Get commits associated with a text pattern

Get commits associated with a text pattern

get_commits_pattern(
  repo = ".",
  pattern = c(Ticket = "#[[:digit:]]+"),
  ref = "master",
  path = NULL,
  silent = FALSE
)

Arguments

repo
a path to a repository or a git_repository object. Default is '.'

pattern
Named vector of regex patterns

ref
The name of a reference to list commits from e.g. a tag or a branch. The default is NULL for the current branch.

path
The path to a file. If not NULL, only commits modifying this file will be returned. Note that modifying commits that occurred before the file was given its present name are not returned; that is, the output of git log with --no-follow is reproduced.

silent
Logical. Whether to hide messages.

Value

A tibble with commits and tags

Examples

repo <- fake_repo()
get_commits_pattern(repo = repo, pattern = c("Ticket" = "#[[:digit:]]+"))

#> 4 commits found.

#> # A tibble: 7 x 11
#>   sha   summary message author email when                order tag.name
#>   <chr> <chr>   <chr>   <chr>  <chr> <dttm>              <int> <chr>   
#> 1 48a1… Add NE… "Add N… Alice  alic… 2020-08-06 13:24:15     4 <NA>    
#> 2 48a1… Add NE… "Add N… Alice  alic… 2020-08-06 13:24:15     4 <NA>    
#> 3 48a1… Add NE… "Add N… Alice  alic… 2020-08-06 13:24:15     4 <NA>    
#> 4 f1ab… Third … "Third… Alice  alic… 2020-08-06 13:24:15     3 v0.1    
#> 5 f1ab… Third … "Third… Alice  alic… 2020-08-06 13:24:15     3 v0.1    
#> 6 aed5… exampl… "examp… Alice  alic… 2020-08-06 13:24:15     2 v0.1    
#> 7 5abe… First … "First… Alice  alic… 2020-08-06 13:24:15     1 v0.1    
#> # … with 3 more variables: tag.message <chr>, pattern.type <chr>,
#> #   pattern.content <chr>

get_commits_pattern(repo = repo,
  pattern = c("Ticket" = "ticket[[:digit:]]+", "Issues" = "#[[:digit:]]+"))

#> 4 commits found.

#> # A tibble: 12 x 11
#>    sha   summary message author email when                order tag.name
#>    <chr> <chr>   <chr>   <chr>  <chr> <dttm>              <int> <chr>   
#>  1 48a1… Add NE… "Add N… Alice  alic… 2020-08-06 13:24:15     4 <NA>    
#>  2 48a1… Add NE… "Add N… Alice  alic… 2020-08-06 13:24:15     4 <NA>    
#>  3 48a1… Add NE… "Add N… Alice  alic… 2020-08-06 13:24:15     4 <NA>    
#>  4 48a1… Add NE… "Add N… Alice  alic… 2020-08-06 13:24:15     4 <NA>    
#>  5 48a1… Add NE… "Add N… Alice  alic… 2020-08-06 13:24:15     4 <NA>    
#>  6 f1ab… Third … "Third… Alice  alic… 2020-08-06 13:24:15     3 v0.1    
#>  7 f1ab… Third … "Third… Alice  alic… 2020-08-06 13:24:15     3 v0.1    
#>  8 f1ab… Third … "Third… Alice  alic… 2020-08-06 13:24:15     3 v0.1    
#>  9 aed5… exampl… "examp… Alice  alic… 2020-08-06 13:24:15     2 v0.1    
#> 10 aed5… exampl… "examp… Alice  alic… 2020-08-06 13:24:15     2 v0.1    
#> 11 5abe… First … "First… Alice  alic… 2020-08-06 13:24:15     1 v0.1    
#> 12 5abe… First … "First… Alice  alic… 2020-08-06 13:24:15     1 v0.1    
#> # … with 3 more variables: tag.message <chr>, pattern.type <chr>,
#> #   pattern.content <chr>