Get commits associated chronologically with tags

Get commits associated chronologically with tags

get_commits_tags(repo = ".", ref = "master", path = NULL, silent = FALSE)

Arguments

repo

a path to a repository or a git_repository object. Default is '.'

ref

The name of a reference to list commits from e.g. a tag or a branch. The default is NULL for the current branch.

path

The path to a file. If not NULL, only commits modifying this file will be returned. Note that modifying commits that occurred before the file was given its present name are not returned; that is, the output of git log with --no-follow is reproduced.

silent

Logical. Whether to hide messages.

Value

A tibble with one line for each commit and the following columns:

sha: sha of the commit

summary: First line of the commit message

message: Full content of the commit message

author: author of the commit

email: email of the author

when: commit time

order: order of commit messages. 1 is the oldest.

tag.name: name of tag associated with all commits since the last tag

tag.message: message of the tagged commit

Examples

repo <- fake_repo()
get_commits_tags(repo = repo)
#> 4 commits found.
#> # A tibble: 4 × 9
#>   sha      summary  message      author email when                order tag.name
#>   <chr>    <chr>    <chr>        <chr>  <chr> <dttm>              <int> <chr>   
#> 1 6fa4832… Add NEWS "Add NEWS\n… Alice  alic… 2022-01-19 08:06:21     4 NA      
#> 2 824ab8a… Third c… "Third comm… Alice  alic… 2022-01-19 08:06:21     3 v0.1    
#> 3 462bd97… example… "example: m… Alice  alic… 2022-01-19 08:06:21     2 v0.1    
#> 4 60c40d5… First c… "First comm… Alice  alic… 2022-01-19 08:06:21     1 v0.1    
#> # … with 1 more variable: tag.message <chr>