git history to bookdown

This function turns the history of git into a bookdown. Each chapter is a group of commits. The first page gives a summary of all the groups.

git_down(repo = ".", book_path = "gitdown", open = TRUE,
  author = "John Doe", pattern, names_section, ref = "master")

Arguments

repo
The path to a repository. Default is `.`

book_path
The path to the bookdown output. Default is `"gitdown"`.

open
Should the bookdown be opened once compiled? Default is TRUE.

author
Author of the Bookdown

pattern
pattern to expose commits, like "#[[:digit:]]" for issues

names_section
names for each section, one pattern = one section

ref
the name of the branch, by default master