Turns the active branch history of git into a bookdown.

Read all commit messages of your local git repository and sort them according to tags or specific text pattern into chapters of a HTML book using 'bookdown'. Each chapter is a group of commits. The first page gives a summary of all the groups.

git_down(
  repo = ".",
  book_path = "gitdown",
  open = TRUE,
  author = "John Doe",
  pattern = c(Issues = "#[[:digit:]]+"),
  pattern.table = NULL,
  ref = "master",
  ...
)

Arguments

repo
The path to a repository. Default is .

book_path
The path to the bookdown output. Default is "gitdown".

open
Should the bookdown be opened once compiled? Default is TRUE.

author
Author of the bookdown

pattern
Named vector with regex pattern to expose commits, like c("Issues" = "#\[\[:digit:\]\]") for issues

pattern.table
data.frame with two columns: pattern and description of the pattern. This is used as correspondence table to add some names to existing patterns.

ref
the name of the branch, by default master

...
Other parameters to pass to rmarkdown::render()

Value

Path of the HTML gitbook saved in the repo/book_path directory.

Examples

repo <- fake_repo()
if (rmarkdown::pandoc_available("1.12.3")) {
  res <- git_down(repo, pattern = c("Tickets" = "ticket[[:digit:]]+", "Issues" = "#[[:digit:]]+"),
    open = FALSE)
}

#> 
#> 
#> processing file: index.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |......................................................................| 100%
#>    inline R code fragments
#> 
#> 

#> output file: index.knit.md

#> /usr/local/bin/pandoc +RTS -K512m -RTS index.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output index.html --lua-filter /Users/runner/work/_temp/Library/bookdown/rmarkdown/lua/custom-environment.lua --lua-filter /Users/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /Users/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/latex-div.lua --metadata-file /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpVNxQdo/file780235652a --wrap preserve --standalone --section-divs --table-of-contents --toc-depth 3 --template /Users/runner/work/_temp/Library/bookdown/templates/gitbook.html --highlight-style pygments --number-sections --css style.css --include-in-header /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpVNxQdo/rmarkdown-str7807f84a987.html --mathjax 

#> 
#> Output created: gitbook-for-git2r-7804e69c5be.html

if (FALSE) {
# Open the book
  browseURL(res)
}
# With table of correspondence
pattern.table <- data.frame(number = c("#2", "#1"),
  title = c("#2 A second issue to illustrate a blog post",
                       "#1 An example of issue"))
if (rmarkdown::pandoc_available("1.12.3")) {
  res <- git_down(repo, pattern = c("Issues" = "#[[:digit:]]+"),
    pattern.table = pattern.table, open = FALSE)
}

#> 
#> 
#> processing file: index.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |......................................................................| 100%
#>    inline R code fragments
#> 
#> 

#> output file: index.knit.md

#> /usr/local/bin/pandoc +RTS -K512m -RTS index.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output index.html --lua-filter /Users/runner/work/_temp/Library/bookdown/rmarkdown/lua/custom-environment.lua --lua-filter /Users/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /Users/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/latex-div.lua --metadata-file /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpVNxQdo/file7805e961f78 --wrap preserve --standalone --section-divs --table-of-contents --toc-depth 3 --template /Users/runner/work/_temp/Library/bookdown/templates/gitbook.html --highlight-style pygments --number-sections --css style.css --include-in-header /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpVNxQdo/rmarkdown-str78051ec2fcb.html --mathjax 

#> 
#> Output created: gitbook-for-git2r-7804e69c5be.html

if (FALSE) {
# Open the book
  browseURL(res)
}