Nest all commits by each pattern found in the commit messages

Nest all commits by each pattern found in the commit messages

nest_commits_by_pattern(
  repo,
  pattern = c(Issues = "#[[:digit:]]+"),
  pattern.table = NULL,
  ref = "master",
  silent = TRUE
)

Arguments

repo
a path to a repository or a git_repository object. Default is '.'

pattern
Named vector with regex pattern to expose commits, like c("Issues" = "#[[:digit:]]") for issues

pattern.table
data.frame with two columns: pattern and description. This is used as correspondence table to add some names to existing patterns.

ref
The name of a reference to list commits from e.g. a tag or a branch. The default is NULL for the current branch.

silent
Logical. Whether to hide messages.

Value

A tibble with a row for each different pattern found and a nested 'data' column with all related commits

Examples

repo <- fake_repo()
nest_commits_by_pattern(repo)

#> # A tibble: 6 x 6
#>   pattern.type pattern.content  pattern.title pattern_numeric link_pattern data 
#>   <chr>        <chr>            <chr>                   <dbl> <chr>        <lis>
#> 1 Issues       #1               #1                          1 issues-1     <tib…
#> 2 Issues       #2               #2                          2 issues-2     <tib…
#> 3 Issues       #12              #12                        12 issues-12    <tib…
#> 4 Issues       #32              #32                        32 issues-32    <tib…
#> 5 Issues       #145             #145                      145 issues-145   <tib…
#> 6 Issues       No related issu… No related i…              NA issues-na    <tib…